moved stuff from folder cms-api to cms directory; deleted unnecessary import in app.module
