+++ Angular/src/app/cms/cms.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CmsGenericPageComponent } from './components/cms-generic-page/cms-generic-page.component';
 import {CmsRoutingModule} from './cms-routing.module';
-import { PiranhaPageComponent } from '../cms-api/piranha-page/piranha-page.component';
-import { BlockComponent } from '../cms-api/cms-shared/block/block.component';
-import { GalleryBlockComponent } from '../cms-api/cms-shared/gallery-block/gallery-block.component';
+import { PiranhaPageComponent } from './components/piranha-page/piranha-page.component';
+import { BlockComponent } from './components/block/block.component';
+import { GalleryBlockComponent } from './components/gallery-block/gallery-block.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
