+++ Angular/src/app/cms/components/block/block.component.less
