+++ Angular/src/app/cms/components/gallery-block/gallery-block.component.html
