+++ Angular/src/app/cms/models/primary-image.model.ts
